// packages/highlight/src/types/DivText.ts
/**
 * A React component to view a PDF document
 *
 * @see https://react-pdf-viewer.dev
 * @license https://www.apache.org/licenses/LICENSE-2.0
 * @copyright 2019-2024 Nguyen Huu Phuoc <me@phuoc.ng>
 */

export interface DivText {
    pageIndex: number;
    divIndex: number;
    textContent: string;
}